Allow half days selection
Change the working days field to a decimal and change all the relevant references.  Add in a new set of calculations to the excel document to help aid testing.

Update the About page to include details on the new version changes + changes that wern't documented for the patch release in the summer.